refactor(file): FileController 분리 및 api 모듈로 이동
